requirements.txt
Django>=1.6
django-contrib-comments>=1.5
django-haystack>=2.4
djangorestframework>=3.3.3
